refactored: changed Long to long

+++ src/main/java/exam/dto/SouvenirDto.java
@@ -5,7 +5,7 @@ import lombok.NonNull;
 
 import java.time.LocalDate;
 
-public record SouvenirDto(Long id, @NonNull String name, @NonNull LocalDate date, double price) {
+public record SouvenirDto(Long id, @NonNull String name, @NonNull LocalDate date, long price) {
     public SouvenirDto {
         if (name.equals("") && price < 0) throw new SouvenirValidationException();
     }
