+++ src/main/java/exam/dto/SouvenirFullDto.java
@@ -4,6 +4,6 @@ import lombok.NonNull;
 
 import java.time.LocalDate;
 
-public record SouvenirFullDto(@NonNull Long id, @NonNull String name, @NonNull LocalDate date, double price,
+public record SouvenirFullDto(@NonNull Long id, @NonNull String name, @NonNull LocalDate date, long price,
                               @NonNull ManufacturerDto manufacturer) {
 }
